refactor: move conversion function to own directory
